+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import './App.css'
-import { Login } from './Login/Login'
+import { Login } from './pages/login/Login'
 
 function App() {
