+++ src/pages/login/Login.module.css
@@ -8,17 +8,22 @@
     justify-content: center;
     align-items: center;
     height: 100vh;
+    width: 100%;
     background: var(--background-color);
-    background-image: url('../assets/backgroudEnter.svg');
+    background-image: url('../../assets/backgroudEnter.webp');
     background-repeat: no-repeat;
-    background-position: right center;
-    background-size: 38% 100%;
+    background-position: left 60%;
+    background-size: 38% 25vw;
+}
+
+.containerLogo{
+    width: 450px;
+    text-align: center;
 }
 
 .containerFilho{
     display: flex;
     flex-direction: column;
-    justify-content: flex-start;
     align-items: center;
     width: 450px;
     height: 300px;
@@ -39,7 +44,7 @@
 
 h2{
     text-align: center;
-    margin-top: 20px;
+    margin-top: 35px;
     margin-bottom: 30px;
     color: var(--text-primary);
 }
@@ -79,12 +84,10 @@ h2{
 
 .inputErro{
     border: 1.5px solid var(--danger);
-
 }
 
 .input:focus{
     border: 2px solid var(--blue-200);
-
 }
 
 .input:hover{
@@ -99,7 +102,7 @@ h2{
 .botoesContainer{
     display: flex;
     flex-direction: column;
-    align-items: flex-start; 
+    align-items: center; 
     width: 360px; 
     gap: 10px;
 }
@@ -129,7 +132,6 @@ h2{
     color: var(--blue-200);
     text-align: left; 
     width: fit-content; 
-    padding-left: 5px; 
 }
 
 .buttonEsqueciASenha:hover{
@@ -137,16 +139,48 @@ h2{
 }
 
 .span{
-    visibility: hidden;
     text-align: left; 
     width: fit-content; 
     padding-left: 5px; 
     border: none;
-}
-
-.spanErro{
     color: var(--danger);
     font-size: 13px;
     visibility: visible;
     font-weight: bold;
+}
+
+@media screen and (max-width: 900px){
+    .container{
+        background-image: none;
+    }
+}
+
+
+@media screen and (max-width: 450px) {
+
+    .container{
+        width: 100vw;
+        background-size: 30% 50%;
+    }
+
+    .containerLogo{
+        width: 100vw;
+    }
+
+    .containerFilho{
+        width: 100vw;
+    }
+
+    .input{
+        width: 80vw;
+    }
+
+    .botoesContainer{
+        width: 100vw;
+    }
+
+    .buttonEntrar{
+        width: 80vw;
+    }
+
 }
